Ensure symbols try to fetch private methods
Earlier, this would fail:

class User < ActiveRecord::Base
  validates :name, verify_array: { in: :valid_names_list }

  private

  def valid_names_list
    ['name1', 'name2']
  end
end

The `private_method_defined?` method can be used to check whether a
private method is defined on the class. So, a method that has a visible
method and a private method can be accessed. If not, we still raise an
error message. This will be modified later to use a much descriptive
message.